requirements.txt
gspread 
oauth2client 
python-gitlab
pandas
twilio
datetime